# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Food Subscription & Ledger Management System API",
    "description": "Production-grade, Clean Architecture REST API for Food Subscription & Ledger Management System.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/customers": {
      "post": {
        "tags": [
          "Customer Management"
        ],
        "summary": "Create a new customer",
        "description": "Creates a new customer profile.\nAutomatically initializes customer with active status.",
        "operationId": "create_customer_api_v1_customers_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CustomerResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Customer Management"
        ],
        "summary": "Search & list customers",
        "description": "Retrieves all active customers, with optional instant filtering by query string.",
        "operationId": "list_customers_api_v1_customers_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term for name, phone, or address",
              "title": "Query"
            },
            "description": "Search term for name, phone, or address"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_CustomerResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/customers/{customer_id}": {
      "get": {
        "tags": [
          "Customer Management"
        ],
        "summary": "Get customer details",
        "description": "Fetches detailed profile information for a single customer.",
        "operationId": "get_customer_api_v1_customers__customer_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Customer ID",
              "title": "Customer Id"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CustomerResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Customer Management"
        ],
        "summary": "Update customer details",
        "description": "Updates editable fields of an existing customer profile.",
        "operationId": "update_customer_api_v1_customers__customer_id__put",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Customer ID",
              "title": "Customer Id"
            },
            "description": "Customer ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CustomerResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Customer Management"
        ],
        "summary": "Archive customer (Soft Delete)",
        "description": "Soft-deletes / archives a customer profile without losing financial or delivery history.",
        "operationId": "archive_customer_api_v1_customers__customer_id__delete",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Customer ID",
              "title": "Customer Id"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CustomerResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/invoices/generate/{subscription_id}": {
      "post": {
        "tags": [
          "Invoice Engine"
        ],
        "summary": "Generate initial invoice",
        "description": "Generates an itemized invoice for a subscription based on selected meals, days, rates, and delivery charges.",
        "operationId": "generate_invoice_api_v1_invoices_generate__subscription_id__post",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Subscription ID",
              "title": "Subscription Id"
            },
            "description": "Subscription ID"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_InvoiceResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoices/recalculate/{subscription_id}": {
      "post": {
        "tags": [
          "Invoice Engine"
        ],
        "summary": "Recalculate invoice after meal cancellations",
        "description": "Recalculates invoice after meal cancellations or early termination.\nComputes consumed meals, unused credits, delivery adjustments, refund due, or pending amount.",
        "operationId": "recalculate_cancellation_invoice_api_v1_invoices_recalculate__subscription_id__post",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Subscription ID",
              "title": "Subscription Id"
            },
            "description": "Subscription ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_InvoiceResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledger/customer/{customer_id}": {
      "get": {
        "tags": [
          "Ledger Engine"
        ],
        "summary": "Get customer ledger statement",
        "description": "Retrieves GPay-style double-entry ledger timeline, active invoice breakdown, payment history, total invoiced, total paid, and current pending balance.",
        "operationId": "get_customer_ledger_api_v1_ledger_customer__customer_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Customer ID",
              "title": "Customer Id"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CustomerLedgerSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/financial": {
      "get": {
        "tags": [
          "Reports Engine"
        ],
        "summary": "Generate financial & operational report",
        "description": "Generates detailed daily, weekly, or monthly financial and meal delivery breakdown reports.",
        "operationId": "get_financial_report_api_v1_reports_financial_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Report start date (YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "Report start date (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Report end date (YYYY-MM-DD)",
              "title": "End Date"
            },
            "description": "Report end date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_FinancialReportSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/export/csv": {
      "get": {
        "tags": [
          "Reports Engine"
        ],
        "summary": "Export report as CSV",
        "description": "Exports report metrics to a downloadable CSV file.",
        "operationId": "export_report_csv_api_v1_reports_export_csv_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Start date",
              "title": "Start Date"
            },
            "description": "Start date"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "End date",
              "title": "End Date"
            },
            "description": "End date"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse_CustomerLedgerSummary_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates whether the request was processed successfully",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerLedgerSummary"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable response message",
            "default": ""
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "description": "List of error messages or validation failures"
          }
        },
        "type": "object",
        "title": "ApiResponse[CustomerLedgerSummary]"
      },
      "ApiResponse_CustomerResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates whether the request was processed successfully",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable response message",
            "default": ""
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "description": "List of error messages or validation failures"
          }
        },
        "type": "object",
        "title": "ApiResponse[CustomerResponse]"
      },
...
      "ApiResponse_FinancialReportSummary_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates whether the request was processed successfully",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialReportSummary"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable response message",
            "default": ""
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "description": "List of error messages or validation failures"
          }
        },
        "type": "object",
        "title": "ApiResponse[FinancialReportSummary]"
      },
...
      "ApiResponse_InvoiceResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates whether the request was processed successfully",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvoiceResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable response message",
            "default": ""
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "description": "List of error messages or validation failures"
          }
        },
        "type": "object",
        "title": "ApiResponse[InvoiceResponse]"
      },
      "ApiResponse_List_CustomerResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates whether the request was processed successfully",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CustomerResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable response message",
            "default": ""
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "description": "List of error messages or validation failures"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[CustomerResponse]]"
      },
...
      "CustomerCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Ravi Kumar"
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "example": "9876543210"
          },
          "address": {
            "type": "string",
            "title": "Address",
            "example": "123 Main Street, Sector 4"
          },
          "landmark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Landmark",
            "example": "Near City Park"
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone",
          "address"
        ],
        "title": "CustomerCreate"
      },
      "CustomerLedgerSummary": {
        "properties": {
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "customer_name": {
            "type": "string",
            "title": "Customer Name"
          },
          "customer_phone": {
            "type": "string",
            "title": "Customer Phone"
          },
          "total_invoiced": {
            "type": "number",
            "title": "Total Invoiced"
          },
          "total_paid": {
            "type": "number",
            "title": "Total Paid"
          },
          "current_balance": {
            "type": "number",
            "title": "Current Balance"
          },
          "payment_status": {
            "type": "string",
            "title": "Payment Status",
            "description": "'paid', 'partially_paid', 'overdue', 'advance'"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LedgerEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "active_invoice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvoiceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "recent_payments": {
            "items": {
              "$ref": "#/components/schemas/PaymentResponse"
            },
            "type": "array",
            "title": "Recent Payments",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "customer_name",
          "customer_phone",
          "total_invoiced",
          "total_paid",
          "current_balance",
          "payment_status",
          "entries"
        ],
        "title": "CustomerLedgerSummary"
      },
      "CustomerResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "example": "Ravi Kumar"
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "example": "9876543210"
          },
          "address": {
            "type": "string",
            "title": "Address",
            "example": "123 Main Street, Sector 4"
          },
          "landmark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Landmark",
            "example": "Near City Park"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/CustomerStatus",
            "default": "active"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone",
          "address",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "CustomerResponse"
      },
      "CustomerStatus": {
        "type": "string",
        "enum": [
          "active",
          "paused",
          "cancelled",
          "archived"
        ],
        "title": "CustomerStatus"
      },
      "CustomerUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "landmark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Landmark"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "CustomerUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvoiceItem": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "number",
            "title": "Unit Price"
          },
          "total_price": {
            "type": "number",
            "title": "Total Price"
          }
        },
        "type": "object",
        "required": [
          "description",
          "quantity",
          "unit_price",
          "total_price"
        ],
        "title": "InvoiceItem"
      },
      "InvoiceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "invoice_number": {
            "type": "string",
            "title": "Invoice Number"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "subscription_id": {
            "type": "string",
            "title": "Subscription Id"
          },
          "customer_name": {
            "type": "string",
            "title": "Customer Name"
          },
          "customer_phone": {
            "type": "string",
            "title": "Customer Phone"
          },
          "billing_date": {
            "type": "string",
            "format": "date",
            "title": "Billing Date"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceItem"
            },
            "type": "array",
            "title": "Items"
          },
          "breakdown": {
            "$ref": "#/components/schemas/InvoiceBreakdown"
          },
          "cancellation_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvoiceCancellationSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "'issued', 'partially_paid', 'paid', 'cancelled', 'adjusted'",
            "default": "issued"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "invoice_number",
          "customer_id",
          "subscription_id",
          "customer_name",
          "customer_phone",
          "billing_date",
          "start_date",
          "end_date",
          "items",
          "breakdown",
          "created_at",
          "updated_at"
        ],
        "title": "InvoiceResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}